move usage limits to separate UsageLimits component
